Composer: allow installation on PHP 8
With the previous version constraints, the WP Test Helper plugin could not be installed on PHP 8. Fixed now.

Note: I haven't tested whether it actually _works_ on PHP 8, but wouldn't expect any big problems and without running it on PHP 8, we'll never find out if there are issues anyway (as there are no tests).