Downgrader: output escape exception arguments
When an exception isn't caught, it is likely to end up on the end-user's screen. While we obviously hope this will never happen, we should make sure that those exception messages are then not subject to XSS attacks, which could be hidden in translations.

This hardens the code in the `Downgrader` class against these type of issues.